pyproject.toml
[project]
name = "piper-onnx"
version = "1.0.6"
description = "Use piper TTS with onnxruntime"
requires-python = ">=3.10"
dependencies = [
    "espeakng-loader>=0.2.4",
    "onnxruntime>=1.20.0",
    "phonemizer-fork>=3.3.2",
    "soundfile>=0.13.1"
]

